Add BMI088 IMU support over SPI, fix accel/gyro read and power-mode bugs

+++ inc/usart.h
@@ -7,7 +7,7 @@
 void usart2_init(void);
 void usart3_init(void);
 void usart2_transmit(uint8_t data);
-void usart2_transmitint(uint8_t value); 
+void usart2_transmitint(int value);
 void usart2_transmitstr(const char *s);
 void usart2_transmitfloat(float value, int decimals);
 void usart3_receive(char buf[100]);
